+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-06T11:09:54.594Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-06T13:25:31.849Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")